[GPEEngine] use a strv as the set_active_plugins() argument.
strv aka const char** seems more often used than glists, because it
makes it easier to know what's the actual content of the list/array
is.

Moreover, GSettings uses strv as the type for string lists, so it seems
sane to use it here as well, as set_active_plugins() is going to be used
with GSettings or GConf like 99% of the time.