[PythonLoader] Ensure libgpe and libgpeui can be imported
This is done by explicitely prepending our pyexecdir into sys.path, so
python can know where to load the bindings from when an app links
against libgpe.

Prepending ensures the current version of the bindings will be picked
if there is an older version available somewhere else in sys.path.