[GPEEngine] Do not load all the plugin loaders.
Before this patch, loaders were all loaded when an unknown loader was
requested. OTOH the actual instantiation of the loader was delayed until
a plugin actually needs it. The downside of this approach is that the
python libraries were linked as soon as any plugin was loaded.

Now, we only try to load the plugin loader which has the right filename
when a plugin requests a loader not available yet, so we won't actually
link to anything unless it becomes necessary. Downside is that we are
now very dependant on the filename, but as we control the loaders (we
dont allow third party loaders for now), it doesn't matter.