[GPEEngine] Change the way search paths are specified, again.
The search paths (formerly named "plugin directories") are now a
construct-time property. It is a simple array of strings, to make it
simpler for libraries to instanciate such an array and free it
afterwards, as the previous approach ended up being quite verbose in
the consumer applications.

There is also extensive doc on the new format for the search path list.