Add a GDestroyNotify to module_register_extension_factory()
The peas_object_module_register_extension() was also renamed
peas_object_module_register_extension_factory() as it is a more explicit
name for a rarely used function.  PeasCreateFunc got renamed accordingly
as well.

The nicety is that if someone used the old API it will now generate a
compilation error.