Allow writing C extensions not inheriting from PeasExtensionBase.
PeasExtensionBase ought to be an optional base class allowing the
extension implementer to get the plugin-info instance and the
information it contains.  The introduction of the "plugin-info"
construct property broke the ability to use GObject (or something else)
as the base class.

We restore this ability by dropping this property from the construct
property vales if the object doesn't have this property.  This
behaviour also makes it possible to use a GInterface to get the
"plugin-info" property.