Adjust to API break in gobject-introspection.
gobject-introspection's libgirepository dropped support for
machine-dependent types (short, int, long) and for time_t, so we need
to adjust as well. This should have no consequencies as according to
the g-i commit the typelib was only using machine-specific types yet.

Also bump the gobject-introspection requirement to ensure we're using
a new enough typelib format.

For more information, see gobject-introspection commit
5cb925b20739c04e18e94a03a3e4e68041894b03: "Don't include
machine-dependent integral types in the typelib"

https://bugzilla.gnome.org/show_bug.cgi?id=624080