Replaced invisible plugins with builtin plugins
"Visible" plugins are replaced by "builtin" ones. The "builtin" value
is read from the plugin info file.

A builtin plugin has the same behaviour as previous invisible plugins:
they cannot be enabled or disabled by the user, and they are hidden in
the plugin manager. The application can still do what it wants with
builtin plugins, as libpeas doesn't handle autoloading them.

It means that totem can drop the "set visibility" code and just load the
plugin straight away when a "builtin" plugin is found. It also means
that anjuta can handle loading and unloading them as required.

A later patch is planned to introduces a "show-builtins" property in
the plugin manager, effectively allowing to show or hide those plugins
without the weird "show-invisible" naming.

https://bugzilla.gnome.org/show_bug.cgi?id=621858