Convert the API using va_list to use GArgument instead.
GArgument is designed to allow passing function arguments around and
processing them quite efficiently, and on the other hand a va_list can
only be processed once and must be used from a single function only.

This proved being error-prone, so this patch just switches all API
playing with function arguments from using va_list to using GArgument.
This means one extra marshalling step in the Seed loader, but that's
all...

This commit also fixes calling methods of the Seed extensions when there
are more than a single (out) argument. It was not working everywhere
because it accessed the va_list content from several distinct functions.