Make PeasExtension implement extension interfaces.
With this commit, we make it possible to use PeasExtension instances
directly with the extension interface methods.

This makes the overal use of libpeas nicer and more typesafe, make it
feel like we're using the real underlying extension instance, and make
it easier to port old code over without changing all the API calls.

https://bugzilla.gnome.org/show_bug.cgi?id=627338