python: Rework refcounting to explicit extension ownership
Previously the python reference was transferred to the Python extension
object. It is more correct to increment the refcount in the extension
and decrement it in the loader, because it's more explicit and won't
break in bizarre way if we create more than one Python extension wrapper
around the same object.