Implemented log hooks in libtesting-util
This allow us to add hooks for warning and criticals that are required
during tests.

These have several advantages over globally allowed patterns in that they
avoid crossover issues when a test expects a warning but another
test's pattern causes it to not be emitted. They are required to be triggered
which avoids old patterns from being kept around when no longer needed.
Most importantly they allow us to test that libpeas handles errors correctly.