Don't use properties to store the extension instance
The properties were intended for the tests but we can just as
easily include the loader's extension.h file and acess it directly.

In the future we will be using get_property() and set_property()
for getting and setting properties on extension instances and if
extensions had their own properties they could shadow the properties
of the instances.