Do not expose the old peas_extension_* API to g-ir-scanner.
g-ir-scanner chokes because it fails to associate the GObject type
returned by the fake _get_type() method with the 'Peas' namespace we use
for this library. So we hide all the peas_extension API as the rest of
this deprecated API is not in a good shape in GI anyway.

Adapted from a patch of Pavel Holejsovsky.

https://bugzilla.gnome.org/show_bug.cgi?id=657586