Support py2 and py3 at the same time
This patch reworks the build system so that we use the same sources to
build both a python2 loader and a python3 loader. The python 2 one
keeps the same name for backward compatibility, while the python 3 one
is explicitely called "python3".
Supporting py2 and py3 in the same build is tricky since the existing
autotools support is limited, so we make some assumptions:
 - there are separate --enable-python2 and --enable-python3 flags,
   --enable-python has been removed
 - the py3 loader is in a separate dir but compliles the files from the
   py2 loader
 - for py2 we only check the binary and the python2-config tool, while
   for py3 we use the proper AM_PATH_PYTHON macro
 - for py2 we manually detect pyexecdir, hopefully with the same logic
   used by the AM_PATH_PYTHON macro
 - we do not check for Python.h header anymore since it is not strictly
   needed and autoconf caching makes it difficult
 - for we build the demos only for python3 (unit tests instead are
   replicated for both)

https://bugzilla.gnome.org/show_bug.cgi?id=691081