engine: use topological sort to resolve dependency graph
Previously, if a plugin info was discovered with dependencies before it's
dependency was discovered, there was a chance that the ordering would be
undefined.

Instead, this uses topological sort and reapplies the sorting after every
addition so that delayed dependencies can be applied.

Since this uses a GList for everything, I switched the GQueue plugin_list
back to a GList *plugin_list.